+++ __tests__/translate.test.ts
@@ -1,5 +1,4 @@
-import { snakeCase } from "lodash";
-
+import snakeCase from "../src/vendor/lodash/snakeCase";
 import { I18n } from "../src/I18n";
 import { Dict, Scope } from "../src/typing";
 import { translations } from "./fixtures/translations";
@@ -28,7 +27,7 @@ test("returns missing message translation for invalid scope", () => {
 test("throws an error if missingBehavior is set to error", () => {
   i18n.missingBehavior = "error";
 
-  expect(() => i18n.t("missing.translation")).toThrowError(
+  expect(() => i18n.t("missing.translation")).toThrow(
     "Missing translation: en.missing.translation",
   );
 });