+++ src/helpers/formatNumber.ts
@@ -1,6 +1,6 @@
 import BigNumber from "bignumber.js";
-import repeat from "lodash/repeat";
 
+import repeat from "../vendor/lodash/repeat";
 import { FormatNumberOptions, Numeric } from "../typing";
 import { roundNumber } from "./roundNumber";