+++ src/helpers/lookup.ts
@@ -1,5 +1,4 @@
-import get from "lodash/get";
-
+import get from "../vendor/lodash/get";
 import { Dict, Scope } from "../typing";
 import { I18n } from "../I18n";
 import { isSet } from "./isSet";