+++ src/helpers/propertyFlatList.ts
@@ -1,7 +1,6 @@
-import isArray from "lodash/isArray";
-import isObject from "lodash/isObject";
-import flattenDeep from "lodash/flattenDeep";
-
+import isArray from "../vendor/lodash/isArray";
+import isObject from "../vendor/lodash/isObject";
+import flattenDeep from "../vendor/lodash/flattenDeep";
 import { Dict } from "../typing";
 
 interface Indexable {